extracting state specific presentation into sub-component.
